requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
crcmod==1.7
cryptography==39.0.2
cssselect==1.2.0
dataclasses-json==0.5.7
dnspython==2.3.0
elevenlabslib==0.4.1
email-validator==1.3.1
et-xmlfile==1.1.0
Flask==2.2.3
Flask-RESTful==0.3.9
frozenlist==1.3.3
google-search-results==2.4.2
googlesearch-python==1.2.3
greenlet==2.0.1
html2text==2020.1.16
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
langchain==0.0.129
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
names==0.3.0
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
oss2==2.17.0
packaging==23.0
playwright==1.31.1
prompt-toolkit==3.0.38
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydantic-computed==0.2.1
pyee==9.0.4
Pygments==2.14.0
pyperclip==1.8.2
pyquery==2.0.0
python-dotenv==1.0.0
python-frontmatter==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rich==13.3.2
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4
SQLAlchemy==1.4.47
tenacity==8.2.2
tiktoken==0.3.2
tqdm==4.65.0
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
Wikipedia-API==0.5.8
yarl==1.8.2
zipp==3.15.0
